cleaning up file structure, combined the two model folders, change example to upper case in the .env.EXAMPLE, removed a schema file that was in the main directory
