+++ models/index.js
@@ -1,6 +1,10 @@
-const todo = require('./todo.js');
-const note = require('./note');
-const calendar = require('./calendar');
+const Todos = require('./Todo')
+const TodoItems = require('./TodoItems')
+const Notes = require('./Notes');
+const NoteItems = require('./NoteItems')
+const Contacts = require('./Contacts')
+const ContactsInfo = require('./ContactsInfo')
+const User = require('./User')
 
 note.hasMany(calendar, {
   foreignKey: 'node_id',
@@ -10,4 +14,4 @@ calendar.belongsTo(note, {
   foreignKey: 'note_id',
 });
 
-module.exports = { todo, note, calendar };
+module.exports = { Todos, TodoItems, Notes, NoteItems, Contacts, ContactsInfo, User };
